#5085 mobile fixes
