+++ styles/global.scss
@@ -270,3 +270,7 @@ hr {
     }
   }
 }
+
+.no-shadow {
+  box-shadow: none !important;
+}
